backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
Pillow==11.1.0
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.6.0
dj-database-url==2.1.0
